Autotracking start: emit foundTask when we match a keyword

+++ src/Task.cpp
@@ -6,15 +6,15 @@
 Task::Task()
 = default;
 
-Task::Task(int tid) {
+Task::Task(qint64 tid) {
     tc_id = tid;
 }
 
-int Task::getTc_id() const {
+qint64 Task::getTc_id() const {
     return tc_id;
 }
 
-void Task::setTc_id(int tc_id) {
+void Task::setTc_id(qint64 tc_id) {
     this->tc_id = tc_id;
 }
