+++ src/Task.h
@@ -6,17 +6,17 @@
 
 class Task {
 private:
-    int tc_id;
+    qint64 tc_id;
     QString name;
     QString keywords;
     QStringList keywordsList;
 
 public:
     Task();
-    Task(int);
+    Task(qint64);
 
-    int getTc_id() const;
-    void setTc_id(int tc_id);
+    qint64 getTc_id() const;
+    void setTc_id(qint64 tc_id);
 
     const QString &getName() const;
     void setName(const QString &name);
